test-support moved to addon-test-support
