+++ .eslintrc.js
@@ -15,6 +15,7 @@ module.exports = {
     browser: true,
   },
   rules: {
+    'no-console': ["error", { allow: ["warn", "error"] }],
   },
   overrides: [
     // node files
