+++ addon-test-support/assertions/assert-tooltip-content.js
@@ -1,0 +1,18 @@
+import { assert as emberAssert } from '@ember/debug';
+import { isNone } from '@ember/utils';
+import { findTooltip } from 'ember-tooltips/test-support';
+
+export default function assertTooltipContent(assert, options = {}) {
+  const { contentString } = options;
+
+  if (isNone(contentString)) {
+    emberAssert('You must specify a contentString property in the options parameter');
+  }
+
+  const $tooltip = findTooltip();
+  const tooltipContent = $tooltip.text().trim();
+
+  assert.equal(tooltipContent, contentString,
+    `Content of tooltip (${tooltipContent}) matched expected (${contentString})`);
+
+}
