+++ addon-test-support/assertions/assert-tooltip-visible.js
@@ -1,0 +1,11 @@
+import { findTooltip } from 'ember-tooltips/test-support';
+
+export default function assertTooltipNotVisible(assert, options = {}) {
+  const { selector } = options;
+  const $tooltip = findTooltip(selector);
+  const ariaHidden = $tooltip.attr('aria-hidden');
+
+  assert.ok(ariaHidden === 'false',
+    `assertTooltipVisible(): the ember-tooltip should be visible:
+      aria-hidden = ${ariaHidden}`);
+}
