+++ addon-test-support/index.js
@@ -1,0 +1,14 @@
+export { default as assertTooltipContent } from './assertions/assert-tooltip-content';
+export { default as assertTooltipNotRendered } from './assertions/assert-tooltip-not-rendered';
+export { default as assertTooltipNotVisible } from './assertions/assert-tooltip-not-visible';
+export { default as assertTooltipRendered } from './assertions/assert-tooltip-rendered';
+export { default as assertTooltipSide } from './assertions/assert-tooltip-side';
+export { default as assertTooltipSpacing } from './assertions/assert-tooltip-spacing';
+export { default as assertTooltipVisible } from './assertions/assert-tooltip-visible';
+
+export { default as findTooltipTarget } from './utils/find-tooltip-target';
+export { default as findTooltip } from './utils/find-tooltip';
+export { default as getOppositeSide } from './utils/get-opposite-side';
+export { default as getPositionDifferences } from './utils/get-position-differences';
+export { default as triggerTooltipTargetEvent } from './utils/trigger-tooltip-target-event';
+export { default as validateSide } from './utils/validate-side';
