+++ addon-test-support/utils/get-opposite-side.js
@@ -1,0 +1,8 @@
+export default function getOppositeSide(side) {
+  switch (side) {
+    case 'top': return 'bottom';
+    case 'right': return 'left';
+    case 'bottom': return 'top';
+    case 'left': return 'right';
+  }
+}
