+++ tests/integration/components/delay-test.js
@@ -9,7 +9,7 @@ import {
   assertTooltipNotVisible,
   triggerTooltipTargetEvent,
   assertTooltipVisible,
-} from 'dummy/tests/helpers/ember-tooltips';
+} from 'ember-tooltips/test-support';
 
 async function testTooltipDelay(assert, template) {
