+++ tests/integration/components/duration-test.js
@@ -8,7 +8,7 @@ import {
   assertTooltipNotVisible,
   assertTooltipVisible,
   triggerTooltipTargetEvent,
-} from 'dummy/tests/helpers/ember-tooltips';
+} from 'ember-tooltips/test-support';
 
 module('Integration | Option | duration', function(hooks) {
   setupRenderingTest(hooks);
