+++ tests/integration/components/event-test.js
@@ -7,7 +7,7 @@ import {
   assertTooltipNotRendered,
   assertTooltipVisible,
   triggerTooltipTargetEvent,
-} from 'dummy/tests/helpers/ember-tooltips';
+} from 'ember-tooltips/test-support';
 
 module('Integration | Option | event', function(hooks) {
   setupRenderingTest(hooks);
