+++ tests/integration/components/helpers/trigger-tooltip-event-test.js
@@ -4,7 +4,7 @@ import { render } from '@ember/test-helpers';
 import hbs from 'htmlbars-inline-precompile';
 import {
   triggerTooltipTargetEvent,
-} from 'dummy/tests/helpers/ember-tooltips';
+} from 'ember-tooltips/test-support';
 
 module('Integration | Helpers | triggerTooltipTargetEvent', function(hooks) {
   setupRenderingTest(hooks);
