+++ tests/integration/components/show-on-test.js
@@ -6,7 +6,7 @@ import {
   assertTooltipNotRendered,
   assertTooltipVisible,
   triggerTooltipTargetEvent,
-} from 'dummy/tests/helpers/ember-tooltips';
+} from 'ember-tooltips/test-support';
 
 module('Integration | Option | showOn', function(hooks) {
   setupRenderingTest(hooks);
