+++ tests/integration/components/target-test.js
@@ -6,7 +6,7 @@ import {
   findTooltip,
   findTooltipTarget,
   triggerTooltipTargetEvent,
-} from 'dummy/tests/helpers/ember-tooltips';
+} from 'ember-tooltips/test-support';
 
 module('Integration | Component | target', function(hooks) {
   setupRenderingTest(hooks);
